requirements.txt
astroid==2.3.3
autopep8==1.4.4
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
csvImporter==0.1.3.2
defusedxml==0.6.0
Django==2.2.7
django-braces==1.13.0
django-cors-middleware==1.4.0
django-extensions==2.2.5
django-freeze==0.6.4
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.10.3
djangorestframework-oauth==1.1.0
drf-base64==2.0
et-xmlfile==1.0.1
idna==2.8
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
openpyxl==3.0.2
# psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
wrapt==1.11.2
xmltodict==0.12.0
